agentic-backed/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0.0
pytest
python-dateutil

# Google ADK and AI dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0
psycopg2-binary
